# payload/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Egregore Projection Plane",
    "version": "0.6.0-phase1"
  },
  "paths": {
...
    "/api/v1/factory": {
      "post": {
        "tags": [
          "factory"
        ],
        "summary": "Run Default Factory",
        "description": "Run the most sophisticated factory mode (coding_factory) by default.",
        "operationId": "run_default_factory_api_v1_factory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactoryRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactoryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/factory/{mode}": {
      "post": {
        "tags": [
          "factory"
        ],
        "summary": "Run Factory",
        "description": "Run the factory pipeline for the selected mode (wheel position).",
        "operationId": "run_factory_api_v1_factory__mode__post",
        "parameters": [
          {
            "name": "mode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactoryRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactoryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FactoryRunRequest": {
        "properties": {
          "input": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 1,
            "title": "Input"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 8192.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "input"
        ],
        "title": "FactoryRunRequest"
      },
      "FactoryRunResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "pipeline_version": {
            "type": "integer",
            "title": "Pipeline Version"
          },
          "final_output": {
            "type": "string",
            "title": "Final Output"
          },
          "stations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StationOutput"
            },
            "type": "object",
            "title": "Stations"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          },
          "gates": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gates"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "pipeline_version",
          "final_output",
          "stations",
          "provenance"
        ],
        "title": "FactoryRunResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StationOutput": {
        "properties": {
          "output": {
            "type": "string",
            "title": "Output"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "tokens": {
            "type": "integer",
            "title": "Tokens",
            "default": 0
          },
          "elapsed_ms": {
            "type": "number",
            "title": "Elapsed Ms",
            "default": 0.0
          },
          "compressed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compressed"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "parsed": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed"
          }
        },
        "type": "object",
        "required": [
          "output",
          "model"
        ],
        "title": "StationOutput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}